add col

+++ database/migrations/2023_05_25_123559_create_trains_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('Arrival_station');
             $table->string('Departure_time');
             $table->string('Arrival_time');
-            $table->tinyInteger('Train_Code') ;
+            $table->smallInteger('Train_Code') ;
             $table->tinyInteger('Number_of_carriages');
             $table->boolean('In_time') ;
             $table->boolean('Deleted');
